pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wayfinders-cli"
version = "0.2.0"
description = "Wayfinders show pipeline CLI scaffold"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12",
  "pydantic>=2.6",
  "pyyaml>=6.0",
  "jinja2>=3.1",
  "rich>=13.7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.5",
]
placeholders = [
  "pillow>=10.0"
]

[project.scripts]
wf = "wayfinders_cli.cli:app"

[tool.ruff]
line-length = 100
target-version = "py310"
